// admisi/src/types/RekapPerubahanProdiTypes.ts
export type RekapPerubahanProdiType = {
  NomorDaftar: string;
  NamaCamhs: string;
  ProgramStudiLama: string;
  ProgramStudiBaru: string;
  ShiftKuliahLama: string;
  ShiftKuliahBaru: string;
  UserValidasi: string;
  WaktuValidasi: string;
  StatusValidasi: string;
  CatatanValidasi: string;
};

export type SejarahTagihanYangDihapusType = {
  NamaProdi: string;
  GelombangNama: string;
  PeriodeNama: string;
  PeringkatNama: string;
  WaktuKuliah: string;
  UserValidasi: string;
  TanggalRekap: string;
};